ai-dev/Text-to-SQL POC v1.0/requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
chromadb>=0.5.0
sqlalchemy>=2.0.0
sentence-transformers>=3.0.0
cx_Oracle>=8.3.0
python-dotenv>=1.0.0
